cities and state search placed on another vm file
